requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
Flask==3.0.3
flatbuffers==24.3.25
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==2.1.1
onnxruntime==1.19.2
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
pyreadline3==3.5.4
sympy==1.13.3
Werkzeug==3.0.4
ultralytics